모델 타입 아이디 long으로 타입변경, 자동 인서트  기능 완성

+++ src/main/java/com/lubid/lubiduser/Repository/DeliveryAddressRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DeliveryAddressRepository extends JpaRepository<DeliveryAddress, Integer> {
+public interface DeliveryAddressRepository extends JpaRepository<DeliveryAddress, Long> {
     // User와 관련된 주소 전부 가져오기
     List<DeliveryAddress> findAllByUser(User user);
